Micro optimize cyclic_difference
Benchmarking CT-INT revealed that this code is incredibly hot, so it is
worthwhile to spend some time micro optimizing it.  First of all,
cyclic_difference is moved to the header to give the compiler a chance to inline
it.  We switch to a branchless sign computation to avoid an unnecessary branch.
Finally we replace the division in the conversion operator to double of tau_t by
multiplication with the precomputed inverse.

Co-authored-by: Nils Wentzell <Wentzell@users.noreply.github.com>